--- RFD/sensor_device/main/sensor_device.c
@@ -1,328 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2021-2022 Espressif Systems (Shanghai) CO LTD
- *
- * SPDX-License-Identifier: CC0-1.0
- *
- * Zigbee HA_TEMPERATURE
- *
- * This example code is in the Public Domain (or CC0 licensed, at your option.)
- *
- * Unless required by applicable law or agreed to in writing, this
- * software is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied.
- */
-#include "sensor_device.h"
-#include "esp_check.h"
-#include "esp_log.h"
-#include "nvs_flash.h"
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include <dht22.h>
-#include "driver/gpio.h"
-#include "battery_read.c"
-#include "zcl/esp_zigbee_zcl_power_config.h"
-#include "ota.c"
-#include "esp_pm.h"
-#include "esp_private/esp_clk.h"
-#include "update_cluster.h"
-
-#if !defined ZB_ED_ROLE
-#error Define ZB_ED_ROLE in idf.py menuconfig to compile RFD (End Device) source code.
-#endif
-
-#if CONFIG_IDF_TARGET_ESP32C6
-#define CONFIG_EXAMPLE_DATA_GPIO GPIO_NUM_2
-#elif CONFIG_IDF_TARGET_ESP32H2
-#define CONFIG_EXAMPLE_DATA_GPIO GPIO_NUM_0
-#endif
-
-#define SENSOR_TYPE DHT_TYPE_AM2301
-
-RTC_DATA_ATTR uint8_t lastBatteryPercentageRemaining = 0x8C;
-
-static char manufacturer[16] = {5, 'B', 'o', 't', 'u', 'k'};
-static char model[16] = {15, 'E', 'S', 'P', '3', '2', 'H', '2', ' ', 'E', 'N', 'D', ' ', 'D', 'e', 'v'};
-static char firmware_version[16] = {6, 'v', 'e', 'r', '0', '.', '5'};
-static const char *TAG = "SENSOR_DEVICE";
-bool connected = false;
-
-
-/********************* Define functions **************************/
-static void bdb_start_top_level_commissioning_cb(uint8_t mode_mask)
-{
-    ESP_ERROR_CHECK(esp_zb_bdb_start_top_level_commissioning(mode_mask));
-}
-
-void measure_temperature()
-{
-    float temperature;
-    uint16_t temperature_to_send = 0;
-
-    float humidity;
-    uint16_t humidity_to_send = 0;
-
-    int battery_level;
-    int battery_voltage;
-    int battery_level_to_send = 0;
-    int battery_voltage_to_send = 0;
-
-    /* Measure temperature loop*/
-    while (1)
-    {
-        if (connected)
-        {
-            if (dht_read_float_data(SENSOR_TYPE, CONFIG_EXAMPLE_DATA_GPIO, &humidity, &temperature) == ESP_OK)
-            {
-                ESP_LOGI(TAG, "Temperature : %.1f ℃", temperature);
-                ESP_LOGI(TAG, "Humidity : %.1f %%", humidity);
-
-                temperature_to_send = (uint16_t)(temperature * 100);
-                humidity_to_send = (uint16_t)(humidity * 100);
-
-                ESP_LOGI(TAG, "Temperature changes, will write new value");
-                zb_update_temp(temperature_to_send, SENSOR_DEVICE_ENDPOINT);
-                ESP_LOGI(TAG, "Humidity changes, will write new value");
-                zb_update_hum(humidity_to_send, SENSOR_DEVICE_ENDPOINT);
-            }
-            else
-            {
-                ESP_LOGW(TAG, "Could not read data from DHT22 Sensor.");
-            }
-
-            if (get_battery_level(&battery_level, &battery_voltage) == ESP_OK)
-            {
-                ESP_LOGI(TAG, "Battery level: %d %%", battery_level);
-                ESP_LOGI(TAG, "Battery voltage: %d mV", battery_voltage);
-                battery_level_to_send = (int)(2 * battery_level);
-                battery_voltage_to_send = (int)(battery_voltage);
-
-                zb_update_battery_level(battery_level_to_send, battery_voltage_to_send, SENSOR_DEVICE_ENDPOINT);
-            }
-            else
-            {
-                ESP_LOGI(TAG, "Could not read battery level and voltage data.");
-            }
-        }
-        else
-        {
-            ESP_LOGI(TAG, "Device is not connected!");
-        }
-        vTaskDelay(pdMS_TO_TICKS(60000));
-    }
-}
-
-void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
-{
-    uint32_t *p_sg_p = signal_struct->p_app_signal;
-    esp_err_t err_status = signal_struct->esp_err_status;
-    esp_zb_app_signal_type_t sig_type = *p_sg_p;
-    esp_zb_zdo_signal_leave_params_t *leave_params = NULL;
-    switch (sig_type)
-    {
-    case ESP_ZB_ZDO_SIGNAL_SKIP_STARTUP:
-        ESP_LOGI(TAG, "Zigbee stack initialized");
-        esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_INITIALIZATION);
-        break;
-    case ESP_ZB_BDB_SIGNAL_DEVICE_FIRST_START:
-    case ESP_ZB_BDB_SIGNAL_DEVICE_REBOOT:
-        if (err_status == ESP_OK)
-        {
-            if (esp_zb_bdb_is_factory_new())
-            {
-                ESP_LOGI(TAG, "Start network steering");
-                esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_NETWORK_STEERING);
-            }
-            else
-            {
-                ESP_LOGI(TAG, "Device restarted");
-            }
-        }
-        else
-        {
-            ESP_LOGW(TAG, "Failed to initialize Zigbee stack (status: %s)", esp_err_to_name(err_status));
-        }
-        break;
-    case ESP_ZB_BDB_SIGNAL_STEERING:
-        if (err_status == ESP_OK)
-        {
-            connected = true;
-            esp_zb_ieee_addr_t extended_pan_id;
-            esp_zb_get_extended_pan_id(extended_pan_id);
-            ESP_LOGI(TAG, "Joined network successfully (Extended PAN ID: %02x:%02x:%02x:%02x:%02x:%02x:%02x:%02x, PAN ID: 0x%04hx, Channel:%d, Short Address: 0x%04hx)",
-                     extended_pan_id[7], extended_pan_id[6], extended_pan_id[5], extended_pan_id[4],
-                     extended_pan_id[3], extended_pan_id[2], extended_pan_id[1], extended_pan_id[0],
-                     esp_zb_get_pan_id(), esp_zb_get_current_channel(), esp_zb_get_short_address());
-        }
-        else
-        {
-            connected = false;
-            ESP_LOGI(TAG, "Network steering was not successful (status: %d)", err_status);
-            esp_zb_scheduler_alarm((esp_zb_callback_t)bdb_start_top_level_commissioning_cb, ESP_ZB_BDB_MODE_NETWORK_STEERING, 1000);
-        }
-        break;
-    case ESP_ZB_ZDO_SIGNAL_LEAVE:
-        leave_params = (esp_zb_zdo_signal_leave_params_t *)esp_zb_app_signal_get_params(p_sg_p);
-        if (leave_params->leave_type == ESP_ZB_NWK_LEAVE_TYPE_RESET)
-        {
-            ESP_LOGI(TAG, "Reset device");
-            esp_zb_factory_reset();
-        }
-        break;
-    case ESP_ZB_COMMON_SIGNAL_CAN_SLEEP:
-        ESP_LOGI(TAG, "Zigbee can sleep");
-        esp_zb_sleep_now();
-        break;
-    default:
-        ESP_LOGI(TAG, "ZDO signal: %s (0x%x), status: %s", esp_zb_zdo_signal_to_string(sig_type), sig_type, esp_err_to_name(err_status));
-        break;
-    }
-}
-
-static esp_err_t zb_action_handler(esp_zb_core_action_callback_id_t callback_id, const void *message)
-{
-    esp_err_t ret = ESP_OK;
-    switch (callback_id)
-    {
-    case ESP_ZB_CORE_OTA_UPGRADE_VALUE_CB_ID:
-        ret = zb_ota_upgrade_status_handler(*(esp_zb_zcl_ota_upgrade_value_message_t *)message);
-        break;
-    default:
-        ESP_LOGW(TAG, "Receive Zigbee action(0x%x) callback", callback_id);
-        break;
-    }
-    return ret;
-}
-
-static esp_err_t esp_zb_power_save_init(void)
-{
-    esp_err_t rc = ESP_OK;
-    int cur_cpu_freq_mhz = CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ;
-    esp_pm_config_t pm_config = {
-        .max_freq_mhz = cur_cpu_freq_mhz,
-        .min_freq_mhz = cur_cpu_freq_mhz,
-        .light_sleep_enable = true};
-    rc = esp_pm_configure(&pm_config);
-    return rc;
-}
-
-static void esp_zb_task(void *pvParameters)
-{
-    /* initialize Zigbee stack */
-    esp_zb_cfg_t zb_nwk_cfg = ESP_ZB_ZED_CONFIG();
-    esp_zb_sleep_enable(true);
-    esp_zb_init(&zb_nwk_cfg);
-    esp_zb_sleep_set_threshold(2000);
-    // TODO: Adjust tx_power for end devices to 0.
-    /* Set trasmitter power tx_power(0) = -24dB */
-    esp_zb_set_tx_power(5);
-
-    uint8_t test_attr;
-    uint8_t power_source = 3;
-    test_attr = 0;
-    /* basic cluster create with fully customized */
-    esp_zb_attribute_list_t *esp_zb_basic_cluster = esp_zb_zcl_attr_list_create(ESP_ZB_ZCL_CLUSTER_ID_BASIC);
-    esp_zb_basic_cluster_add_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_MANUFACTURER_NAME_ID, manufacturer);
-    esp_zb_basic_cluster_add_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_MODEL_IDENTIFIER_ID, model);
-    esp_zb_basic_cluster_add_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_SW_BUILD_ID, firmware_version);
-    esp_zb_basic_cluster_add_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_ZCL_VERSION_ID, &test_attr);
-    esp_zb_basic_cluster_add_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_POWER_SOURCE_ID, &power_source);
-    esp_zb_cluster_update_attr(esp_zb_basic_cluster, ESP_ZB_ZCL_ATTR_BASIC_ZCL_VERSION_ID, &test_attr);
-
-    /* identify cluster create with fully customized */
-    esp_zb_attribute_list_t *esp_zb_identify_cluster = esp_zb_zcl_attr_list_create(ESP_ZB_ZCL_CLUSTER_ID_IDENTIFY);
-    esp_zb_identify_cluster_add_attr(esp_zb_identify_cluster, ESP_ZB_ZCL_ATTR_IDENTIFY_IDENTIFY_TIME_ID, &test_attr);
-    /* create client role of the cluster */
-
-    uint16_t undefined_value;
-    undefined_value = 0x8000;
-    /* Temperature cluster */
-    uint16_t temperature_max = 50000;
-    uint16_t temperature_min = -1000;
-    esp_zb_attribute_list_t *esp_zb_temperature_meas_cluster = esp_zb_zcl_attr_list_create(ESP_ZB_ZCL_CLUSTER_ID_TEMP_MEASUREMENT);
-    esp_zb_temperature_meas_cluster_add_attr(esp_zb_temperature_meas_cluster, ESP_ZB_ZCL_ATTR_TEMP_MEASUREMENT_VALUE_ID, &undefined_value);
-    esp_zb_temperature_meas_cluster_add_attr(esp_zb_temperature_meas_cluster, ESP_ZB_ZCL_ATTR_TEMP_MEASUREMENT_MIN_VALUE_ID, &temperature_min);
-    esp_zb_temperature_meas_cluster_add_attr(esp_zb_temperature_meas_cluster, ESP_ZB_ZCL_ATTR_TEMP_MEASUREMENT_MAX_VALUE_ID, &temperature_max);
-
-    /* Humidity cluster */
-    uint32_t humidity_max = 100000;
-    uint16_t humidity_min = 0;
-    esp_zb_attribute_list_t *esp_zb_hum_meas_cluster = esp_zb_zcl_attr_list_create(ESP_ZB_ZCL_CLUSTER_ID_REL_HUMIDITY_MEASUREMENT);
-    esp_zb_humidity_meas_cluster_add_attr(esp_zb_hum_meas_cluster, ESP_ZB_ZCL_ATTR_REL_HUMIDITY_MEASUREMENT_VALUE_ID, &undefined_value);
-    esp_zb_humidity_meas_cluster_add_attr(esp_zb_hum_meas_cluster, ESP_ZB_ZCL_ATTR_REL_HUMIDITY_MEASUREMENT_MIN_VALUE_ID, &humidity_min);
-    esp_zb_humidity_meas_cluster_add_attr(esp_zb_hum_meas_cluster, ESP_ZB_ZCL_ATTR_REL_HUMIDITY_MEASUREMENT_MAX_VALUE_ID, &humidity_max);
-
-    // TODO: add power_cluster for battery
-    /* POWER_CONFIG cluster */
-    esp_zb_power_config_cluster_cfg_t power_cfg = {0};
-    uint8_t batteryRatedVoltage = 90;
-    uint8_t batteryMinVoltage = 70;
-    uint8_t batteryQuantity = 1;
-    uint8_t batterySize = 0x02;
-    uint16_t batteryAhrRating = 50000;
-    uint8_t batteryAlarmMask = 0;
-    uint8_t batteryVoltage = 90;
-    esp_zb_attribute_list_t *esp_zb_power_cluster = esp_zb_power_config_cluster_create(&power_cfg);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_ID, &batteryVoltage);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_SIZE_ID, &batterySize);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_QUANTITY_ID, &batteryQuantity);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_RATED_VOLTAGE_ID, &batteryRatedVoltage);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_ALARM_MASK_ID, &batteryAlarmMask);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_VOLTAGE_MIN_THRESHOLD_ID, &batteryMinVoltage);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_A_HR_RATING_ID, &batteryAhrRating);
-    esp_zb_power_config_cluster_add_attr(esp_zb_power_cluster, ESP_ZB_ZCL_ATTR_POWER_CONFIG_BATTERY_PERCENTAGE_REMAINING_ID, &lastBatteryPercentageRemaining);
-    // remove 8 first cluster: shift the pointer to the 8th cluster
-    for (int i = 0; i < 7; i++)
-    {
-        esp_zb_power_cluster = esp_zb_power_cluster->next;
-    }
-
-    /* OTA Upgrade Cluster*/
-    esp_zb_ota_cluster_cfg_t ota_cluster_cfg = {
-        .ota_upgrade_file_version = OTA_UPGRADE_RUNNING_FILE_VERSION,
-        .ota_upgrade_downloaded_file_ver = OTA_UPGRADE_DOWNLOADED_FILE_VERSION,
-        .ota_upgrade_manufacturer = OTA_UPGRADE_MANUFACTURER,
-        .ota_upgrade_image_type = OTA_UPGRADE_IMAGE_TYPE,
-    };
-    esp_zb_attribute_list_t *esp_zb_ota_client_cluster = esp_zb_ota_cluster_create(&ota_cluster_cfg);
-    /** add client parameters to ota client cluster */
-    esp_zb_zcl_ota_upgrade_client_variable_t variable_config = {
-        .timer_query = ESP_ZB_ZCL_OTA_UPGRADE_QUERY_TIMER_COUNT_DEF,
-        .hw_version = OTA_UPGRADE_HW_VERSION,
-        .max_data_size = OTA_UPGRADE_MAX_DATA_SIZE,
-    };
-    esp_zb_ota_cluster_add_attr(esp_zb_ota_client_cluster, ESP_ZB_ZCL_ATTR_OTA_UPGRADE_CLIENT_DATA_ID, (void *)&variable_config);
-
-    /* create cluster lists for this endpoint */
-    esp_zb_cluster_list_t *esp_zb_cluster_list = esp_zb_zcl_cluster_list_create();
-    esp_zb_cluster_list_add_basic_cluster(esp_zb_cluster_list, esp_zb_basic_cluster, ESP_ZB_ZCL_CLUSTER_SERVER_ROLE);
-    esp_zb_cluster_list_add_identify_cluster(esp_zb_cluster_list, esp_zb_identify_cluster, ESP_ZB_ZCL_CLUSTER_SERVER_ROLE);
-    esp_zb_cluster_list_add_temperature_meas_cluster(esp_zb_cluster_list, esp_zb_temperature_meas_cluster, ESP_ZB_ZCL_CLUSTER_SERVER_ROLE);
-    esp_zb_cluster_list_add_humidity_meas_cluster(esp_zb_cluster_list, esp_zb_hum_meas_cluster, ESP_ZB_ZCL_CLUSTER_SERVER_ROLE);
-    esp_zb_cluster_list_add_power_config_cluster(esp_zb_cluster_list, esp_zb_power_cluster, ESP_ZB_ZCL_CLUSTER_SERVER_ROLE);
-    esp_zb_cluster_list_add_ota_cluster(esp_zb_cluster_list, esp_zb_ota_client_cluster, ESP_ZB_ZCL_CLUSTER_CLIENT_ROLE);
-
-    esp_zb_ep_list_t *esp_zb_ep_list = esp_zb_ep_list_create();
-    esp_zb_ep_list_add_ep(esp_zb_ep_list, esp_zb_cluster_list, SENSOR_DEVICE_ENDPOINT, ESP_ZB_AF_HA_PROFILE_ID, ESP_ZB_HA_TEMPERATURE_SENSOR_DEVICE_ID);
-
-    esp_zb_device_register(esp_zb_ep_list);
-    esp_zb_core_action_handler_register(zb_action_handler);
-    esp_zb_set_primary_network_channel_set(ESP_ZB_PRIMARY_CHANNEL_MASK);
-    ESP_ERROR_CHECK(esp_zb_start(false));
-    esp_zb_main_loop_iteration();
-}
-
-void app_main(void)
-{
-    esp_zb_platform_config_t config = {
-        .radio_config = ESP_ZB_DEFAULT_RADIO_CONFIG(),
-        .host_config = ESP_ZB_DEFAULT_HOST_CONFIG(),
-    };
-
-    ESP_ERROR_CHECK(voltage_calculate_init());
-
-    ESP_ERROR_CHECK(nvs_flash_init());
-    ESP_ERROR_CHECK(esp_zb_platform_config(&config));
-    ESP_ERROR_CHECK(esp_zb_power_save_init());
-    xTaskCreate(measure_temperature, "measure_temperature", configMINIMAL_STACK_SIZE * 3, NULL, 5, NULL);
-    xTaskCreate(esp_zb_task, "Zigbee_main", 4096, NULL, 6, NULL);
-}
