+++ RFD/sensor_device_deep_sleep/CMakeLists.txt
@@ -14,5 +14,5 @@ set(EXTRA_COMPONENT_DIRS
     ${CMAKE_CURRENT_SOURCE_DIR}/../../components/create_cluster
     )
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(sensor_device)
+project(sensor_device_deep_sleep)
