+++ RFD/sensor_device_light_sleep/main/CMakeLists.txt
