+++ RFD/sensor_device_light_sleep/main/sensor_device.c
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: CC0-1.0
  *
- * Zigbee Sensor-Device
+ * Zigbee HA_TEMPERATURE
  *
  * This example code is in the Public Domain (or CC0 licensed, at your option.)
  *
@@ -16,11 +16,13 @@
 #include "esp_check.h"
 #include "esp_log.h"
 #include "dht22.h"
-#include "battery_read.c"
 #include "ota.c"
-#include "deep_sleep.c"
 #include "update_cluster.h"
 #include "create_cluster.h"
+#ifdef CONFIG_PM_ENABLE
+#include "esp_pm.h"
+#include "esp_private/esp_clk.h"
+#endif
 
 #if !defined ZB_ED_ROLE
 #error Define ZB_ED_ROLE in idf.py menuconfig to compile RFD (End Device) source code.
@@ -36,6 +38,8 @@
 
 static char firmware_version[16] = {7, 'v', 'e', 'r', '0', '.', '1', '0'};
 static const char *TAG = "SENSOR_DEVICE";
+bool connected = false;
+
 
 /********************* Define functions **************************/
 static void bdb_start_top_level_commissioning_cb(uint8_t mode_mask)
@@ -46,39 +50,50 @@ static void bdb_start_top_level_commissioning_cb(uint8_t mode_mask)
 void measure_temperature()
 {
     float temperature;
+    uint16_t temperature_to_send = 0;
+
     float humidity;
-    uint8_t battery_level;
-    uint8_t battery_voltage;
+    uint16_t humidity_to_send = 0;
 
     /* Measure temperature loop*/
-    if (dht_read_float_data(SENSOR_TYPE, CONFIG_EXAMPLE_DATA_GPIO, &humidity, &temperature) == ESP_OK)
+    while (1)
     {
-        ESP_LOGI(TAG, "Temperature : %.1f ℃", temperature);
-        ESP_LOGI(TAG, "Humidity : %.1f %%", humidity);
-        zb_update_temp((int16_t)(temperature * 100), SENSOR_DEVICE_ENDPOINT);
-        zb_update_hum((uint16_t)(humidity * 100), SENSOR_DEVICE_ENDPOINT);
-    }
-    else
-    {
-        ESP_LOGW(TAG, "Could not read data from DHT22 Sensor.");
-    }
+        if (connected)
+        {
+            if (dht_read_float_data(SENSOR_TYPE, CONFIG_EXAMPLE_DATA_GPIO, &humidity, &temperature) == ESP_OK)
+            {
+                ESP_LOGI(TAG, "Temperature : %.1f ℃", temperature);
+                ESP_LOGI(TAG, "Humidity : %.1f %%", humidity);
 
-    if (get_battery_level(&battery_level, &battery_voltage) == ESP_OK) {
-        ESP_LOGI(TAG, "Battery level: %d %%", battery_level);
-        ESP_LOGI(TAG, "Battery voltage: %d mV", battery_voltage);
-        zb_update_battery_level((uint8_t)(2 * battery_level), (uint8_t)(battery_voltage), SENSOR_DEVICE_ENDPOINT);
-    } else {
-        ESP_LOGI(TAG, "Could not read battery level and voltage data.");
+                temperature_to_send = (uint16_t)(temperature * 100);
+                humidity_to_send = (uint16_t)(humidity * 100);
+
+                ESP_LOGI(TAG, "Temperature changes, will write new value");
+                zb_update_temp(temperature_to_send, SENSOR_DEVICE_ENDPOINT);
+                ESP_LOGI(TAG, "Humidity changes, will write new value");
+                zb_update_hum(humidity_to_send, SENSOR_DEVICE_ENDPOINT);
+            }
+            else
+            {
+                ESP_LOGW(TAG, "Could not read data from DHT22 Sensor.");
+            }
+        }
+        else
+        {
+            ESP_LOGI(TAG, "Device is not connected!");
+        }
+        vTaskDelay(pdMS_TO_TICKS(60000));
     }
 }
 
 void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
 {
-    uint32_t *p_sg_p       = signal_struct->p_app_signal;
+    uint32_t *p_sg_p = signal_struct->p_app_signal;
     esp_err_t err_status = signal_struct->esp_err_status;
     esp_zb_app_signal_type_t sig_type = *p_sg_p;
     esp_zb_zdo_signal_leave_params_t *leave_params = NULL;
-    switch (sig_type) {
+    switch (sig_type)
+    {
     case ESP_ZB_ZDO_SIGNAL_SKIP_STARTUP:
         ESP_LOGI(TAG, "Zigbee stack initialized");
         esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_INITIALIZATION);
@@ -87,45 +102,50 @@ void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
     case ESP_ZB_BDB_SIGNAL_DEVICE_REBOOT:
         if (err_status == ESP_OK)
         {
-            ESP_LOGI(TAG, "Start network steering");
-            esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_NETWORK_STEERING);
+            if (esp_zb_bdb_is_factory_new())
+            {
+                ESP_LOGI(TAG, "Start network steering");
+                esp_zb_bdb_start_top_level_commissioning(ESP_ZB_BDB_MODE_NETWORK_STEERING);
+            }
+            else
+            {
+                ESP_LOGI(TAG, "Device restarted");
+            }
         }
         else
         {
-            /* commissioning failed */
-            ESP_LOGW(TAG, "Failed to initialize Zigbee stack (status: %d)", err_status);
-            esp_zb_scheduler_alarm((esp_zb_callback_t)bdb_start_top_level_commissioning_cb, ESP_ZB_BDB_MODE_NETWORK_STEERING, 1000);
+            ESP_LOGW(TAG, "Failed to initialize Zigbee stack (status: %s)", esp_err_to_name(err_status));
         }
         break;
     case ESP_ZB_BDB_SIGNAL_STEERING:
-        if (err_status == ESP_OK) {
+        if (err_status == ESP_OK)
+        {
+            connected = true;
             esp_zb_ieee_addr_t extended_pan_id;
             esp_zb_get_extended_pan_id(extended_pan_id);
             ESP_LOGI(TAG, "Joined network successfully (Extended PAN ID: %02x:%02x:%02x:%02x:%02x:%02x:%02x:%02x, PAN ID: 0x%04hx, Channel:%d, Short Address: 0x%04hx)",
                      extended_pan_id[7], extended_pan_id[6], extended_pan_id[5], extended_pan_id[4],
                      extended_pan_id[3], extended_pan_id[2], extended_pan_id[1], extended_pan_id[0],
                      esp_zb_get_pan_id(), esp_zb_get_current_channel(), esp_zb_get_short_address());
-            
-            /* Start the one-shot timer */
-            const int before_deep_sleep_time_sec = 5;
-            ESP_LOGI(TAG, "Start one-shot timer for %ds to enter the deep sleep", before_deep_sleep_time_sec);
-            measure_temperature();
-            // TODO: For OTA update will be need a if condition.
-            ESP_ERROR_CHECK(esp_timer_start_once(s_oneshot_timer, before_deep_sleep_time_sec * 1000000));
-        } else {
+        }
+        else
+        {
+            connected = false;
             ESP_LOGI(TAG, "Network steering was not successful (status: %d)", err_status);
             esp_zb_scheduler_alarm((esp_zb_callback_t)bdb_start_top_level_commissioning_cb, ESP_ZB_BDB_MODE_NETWORK_STEERING, 1000);
         }
         break;
     case ESP_ZB_ZDO_SIGNAL_LEAVE:
         leave_params = (esp_zb_zdo_signal_leave_params_t *)esp_zb_app_signal_get_params(p_sg_p);
-        if (leave_params->leave_type == ESP_ZB_NWK_LEAVE_TYPE_RESET) {
+        if (leave_params->leave_type == ESP_ZB_NWK_LEAVE_TYPE_RESET)
+        {
             ESP_LOGI(TAG, "Reset device");
             esp_zb_factory_reset();
         }
         break;
-      case ESP_ZB_COMMON_SIGNAL_CAN_SLEEP:
+    case ESP_ZB_COMMON_SIGNAL_CAN_SLEEP:
         ESP_LOGI(TAG, "Zigbee can sleep");
+        esp_zb_sleep_now();
         break;
     default:
         ESP_LOGI(TAG, "ZDO signal: %s (0x%x), status: %s", esp_zb_zdo_signal_to_string(sig_type), sig_type, esp_err_to_name(err_status));
@@ -136,7 +156,8 @@ void esp_zb_app_signal_handler(esp_zb_app_signal_t *signal_struct)
 static esp_err_t zb_action_handler(esp_zb_core_action_callback_id_t callback_id, const void *message)
 {
     esp_err_t ret = ESP_OK;
-    switch (callback_id) {
+    switch (callback_id)
+    {
     case ESP_ZB_CORE_OTA_UPGRADE_VALUE_CB_ID:
         ret = zb_ota_upgrade_status_handler(*(esp_zb_zcl_ota_upgrade_value_message_t *)message);
         break;
@@ -147,12 +168,28 @@ static esp_err_t zb_action_handler(esp_zb_core_action_callback_id_t callback_id,
     return ret;
 }
 
+static esp_err_t esp_zb_power_save_init(void)
+{
+    esp_err_t rc = ESP_OK;
+    int cur_cpu_freq_mhz = CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ;
+    esp_pm_config_t pm_config = {
+        .max_freq_mhz = cur_cpu_freq_mhz,
+        .min_freq_mhz = cur_cpu_freq_mhz,
+        .light_sleep_enable = true};
+    rc = esp_pm_configure(&pm_config);
+    return rc;
+}
+
 static void esp_zb_task(void *pvParameters)
 {
     /* initialize Zigbee stack */
     esp_zb_cfg_t zb_nwk_cfg = ESP_ZB_ZED_CONFIG();
+    esp_zb_sleep_enable(true);
     esp_zb_init(&zb_nwk_cfg);
-    esp_zb_set_tx_power(TX_POWER);
+    esp_zb_sleep_set_threshold(4000);
+    // TODO: Adjust tx_power for end devices to 0.
+    /* Set trasmitter power tx_power(0) = -24dB */
+    esp_zb_set_tx_power(5);
 
     /* create cluster lists for this endpoint */
     esp_zb_cluster_list_t *esp_zb_cluster_list = esp_zb_zcl_cluster_list_create();
@@ -161,13 +198,13 @@ static void esp_zb_task(void *pvParameters)
     create_identify_cluster(esp_zb_cluster_list);
     create_temp_cluster(esp_zb_cluster_list);
     create_hum_cluster(esp_zb_cluster_list);
-    create_battery_cluster(esp_zb_cluster_list);
     create_ota_cluster(esp_zb_cluster_list,
-                        OTA_UPGRADE_RUNNING_FILE_VERSION,
-                        OTA_UPGRADE_DOWNLOADED_FILE_VERSION,
-                        OTA_UPGRADE_MANUFACTURER,
-                        OTA_UPGRADE_IMAGE_TYPE,
-                        OTA_UPGRADE_HW_VERSION);
+                       OTA_UPGRADE_RUNNING_FILE_VERSION,
+                       OTA_UPGRADE_DOWNLOADED_FILE_VERSION,
+                       OTA_UPGRADE_MANUFACTURER,
+                       OTA_UPGRADE_IMAGE_TYPE,
+                       OTA_UPGRADE_HW_VERSION);
+
 
     esp_zb_ep_list_t *esp_zb_ep_list = esp_zb_ep_list_create();
     esp_zb_ep_list_add_ep(esp_zb_ep_list, esp_zb_cluster_list, SENSOR_DEVICE_ENDPOINT, ESP_ZB_AF_HA_PROFILE_ID, ESP_ZB_HA_TEMPERATURE_SENSOR_DEVICE_ID);
@@ -179,15 +216,16 @@ static void esp_zb_task(void *pvParameters)
     esp_zb_main_loop_iteration();
 }
 
-void app_main(void) {
+void app_main(void)
+{
     esp_zb_platform_config_t config = {
         .radio_config = ESP_ZB_DEFAULT_RADIO_CONFIG(),
         .host_config = ESP_ZB_DEFAULT_HOST_CONFIG(),
     };
 
-    ESP_ERROR_CHECK(voltage_calculate_init());
     ESP_ERROR_CHECK(nvs_flash_init());
     ESP_ERROR_CHECK(esp_zb_platform_config(&config));
-    zb_deep_sleep_init();
+    ESP_ERROR_CHECK(esp_zb_power_save_init());
+    xTaskCreate(measure_temperature, "measure_temperature", configMINIMAL_STACK_SIZE * 3, NULL, 5, NULL);
     xTaskCreate(esp_zb_task, "Zigbee_main", 4096, NULL, 6, NULL);
 }
