+++ RFD/sensor_device_light_sleep/partitions.csv
