+++ RFD/sensor_device_light_sleep/sdkconfig.defaults
